requirements.txt
drain3>=0.9.7
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
transformers>=4.38.0
accelerate>=0.16.0
colorama>=0.4.4
tqdm>=4.62.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
aiohttp>=3.8.0 